refactor: move Opus approval logic to separate file
Extract askOpusApproval and related helpers to opus-approval.ts
for better code organization.
